# Conflicts:
#	src/main/java/com/codegym/bemd4/config/SecurityConfiguration.java
#	src/main/java/com/codegym/bemd4/controller/UserController.java
#	src/main/java/com/codegym/bemd4/model/service/impl/LandlordServiceImpl.java
#	src/main/java/com/codegym/bemd4/model/service/impl/UserServiceImpl.java

+++ src/main/java/com/codegym/bemd4/model/entity/person/Landlord.java
@@ -20,11 +20,11 @@ import java.util.Set;
 @Entity
 @Table
 @NoArgsConstructor @AllArgsConstructor @Getter @Setter
-public class Landlord implements UserLoginDetails {
+public class                Landlord implements UserLoginDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @ManyToMany(fetch = FetchType.LAZY)
     @JoinTable(name = "landlord_roles",
